end 1st day

+++ src/main/groovy/eu/solidcraft/starter/examples/task1/rule/ScoringRule.java
@@ -1,6 +1,8 @@
 package eu.solidcraft.starter.examples.task1.rule;
 
+import java.math.BigDecimal;
+
 public interface ScoringRule {
 
-    Long getScore();
+    BigDecimal getScore();
 }
